modularize a bit further and change icons and titles
