+++ beydamaxapp/app/producto/[slug]/page.jsx
@@ -1,4 +1,3 @@
-import NavBar from "@/app/components/NavBar";
 import Header from "./components/Header";
 import ProductNavBar from "./components/ProductNavBar";
 import Title from "./components/Title";
@@ -10,24 +9,18 @@ import ReservationCard from "./components/ReservationCard";
 
 export default function DetalleProducto() {
   return(
-    <main className="bg-gray-100 min-h-screen w-screen">
-      <main className="max-w-screen-2xl m-auto bg-white">
-        < NavBar />
-        < Header />
-        <div className="flex m-auto w-2/3 justify-between items-start 0 -mt-11">
-          <div className="bg-white w-[70%] rounded p-3 shadow">
-            < ProductNavBar />
-            < Title />
-            < Rating />
-            < Description />
-            < Images />
-            < Reviews />
-          </div>
-          <div className="w-[27%] relative text-reg">
-            < ReservationCard />
-          </div>
-        </div>
-      </main>
-    </main>
+    <>
+      <div className="bg-white w-[70%] rounded p-3 shadow">
+        < ProductNavBar />
+        < Title />
+        < Rating />
+        < Description />
+        < Images />
+        < Reviews />
+      </div>
+      <div className="w-[27%] relative text-reg">
+        < ReservationCard />
+      </div>
+    </>
   )
 }
