+++ beydamaxapp/app/search/page.jsx
@@ -1,21 +1,21 @@
-import NavBar from "../components/NavBar";
 import Header from "./components/Header";
 import SearchSidebar from "./components/SearchSidebar";
 import ProductCard from "./components/ProductCard";
 
+export const metadata = {
+  title: 'Search | Milesstone grill',
+}
+
 export default function Search() {
   return(
-    <main className="bg-gray-100 min-h-screen w-screen">
-      <main className="max-w-screen-2xl m-auto bg-white">
-        < NavBar />
-        < Header />
-        <div className="flex py-4 m-auto w-2/3 justify-between items-start">
-          < SearchSidebar />
-          <div className="w-5/6">
-            < ProductCard />
-          </div>
+    <>
+      < Header />
+      <div className="flex py-4 m-auto w-2/3 justify-between items-start">
+        < SearchSidebar />
+        <div className="w-5/6">
+          < ProductCard />
         </div>
-      </main>
-    </main>
+      </div>
+    </>
   )
 }
